Move AI messages to a new file
